Rename files so all test files are detected and ran under all django versions
