+++ django_ses/tests/__init__.py
@@ -1,6 +1,6 @@
-from .backend import *
-from .commands import *
-from .settings import *
-from .stats import *
-from .views import *
+from .test_backend import *
+from .test_commands import *
+from .test_settings import *
+from .test_stats import *
+from .test_views import *
 from .test_verifier import *
