+++ django_ses/tests/test_settings.py
