+++ django_ses/tests/test_stats.py
