+++ django_ses/tests/test_views.py
